changed price datatype float to decimal

+++ server/api/migrations/20201125221836-create-item.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(10, 2),
       },
       createdAt: {
         allowNull: false,
